+++ src/Compze.Tessaging/Implementation/DistributedTeventStoreTeventPublisher.cs
@@ -25,10 +25,10 @@ static class DistributedTeventStoreTeventPublisherRegistrar
    readonly IOutbox _outbox = outbox;
    readonly IInProcessTeventPublisher _inProcessTeventPublisher = inProcessTeventPublisher;
 
-   void ITeventStoreTeventPublisher.Publish(ITaggregateTevent tevent, IScopeResolver scopeResolver)
+   void ITeventStoreTeventPublisher.Publish(ITaggregateIdentifyingTevent<ITaggregateTevent> wrappedTevent, IScopeResolver scopeResolver)
    {
-      TessageInspector.AssertValidToSendRemote(tevent);
-      _inProcessTeventPublisher.Publish(tevent, scopeResolver);
-      _outbox.PublishTransactionally(tevent);
+      TessageInspector.AssertValidToSendRemote(wrappedTevent);
+      _inProcessTeventPublisher.Publish(wrappedTevent, scopeResolver);
+      _outbox.PublishTransactionally(wrappedTevent);
    }
 }